move code to the location module
