+++ urbanairship-location-react-native/android/src/main/java/com/urbanairship/reactnative/location/AirshipLocationPackage.java
@@ -13,7 +13,7 @@ import com.facebook.react.bridge.JavaScriptModule;
 public class AirshipLocationPackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        return Arrays.<NativeModule>asList(new AirshipLocationModule(reactContext));
+        return Arrays.<NativeModule>asList(new AirshipLocationReactModule(reactContext));
     }
 
     @Override
