+++ urbanairship-location-react-native/android/src/main/java/com/urbanairship/reactnative/location/AirshipLocationReactModule.java
@@ -1,10 +1,18 @@
 package com.urbanairship.reactnative.location;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
+import android.os.AsyncTask;
+import android.os.Build;
+
+import androidx.core.content.ContextCompat;
+
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.Callback;
 import com.urbanairship.location.AirshipLocationManager;
+import com.facebook.react.bridge.Promise;
 
 public class AirshipLocationReactModule extends ReactContextBaseJavaModule {
 
@@ -72,4 +80,17 @@ public class AirshipLocationReactModule extends ReactContextBaseJavaModule {
         promise.resolve(AirshipLocationManager.shared().isBackgroundLocationAllowed());
     }
 
+    /**
+     * Helper to determine if location permissions should be requested or not.
+     *
+     * @return {@code true} if permissions should be requested, otherwise {@code false}.
+     */
+    private boolean shouldRequestLocationPermissions() {
+        if (Build.VERSION.SDK_INT < 23) {
+            return false;
+        }
+
+        return ContextCompat.checkSelfPermission(getReactApplicationContext(), Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_DENIED &&
+                ContextCompat.checkSelfPermission(getReactApplicationContext(), Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_DENIED;
+    }
 }
