+++ urbanairship-location-react-native/ios/AirshipLocationReactModule.h
@@ -1,5 +1,7 @@
 /* Copyright Airship and Contributors */
 
+#import <React/RCTEventEmitter.h>
+
 #if __has_include("AirshipLib.h")
 #import "AirshipLib.h"
 #else
