+++ urbanairship-location-react-native/ios/AirshipLocationReactModule.m
@@ -4,7 +4,6 @@
 
 @implementation AirshipLocationReactModule
 
-
 #pragma mark -
 #pragma mark Module methods
